pyproject.toml
[project]
name = "NodeRAG"
version = "0.2.0"
description = "NodeRAG, a graph-centric framework introducing heterogeneous graph structures that enable the seamless and holistic integration of graph-based methodologies into the RAG workflow"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.13",
    "aiosignal==1.3.2",
    "altair==5.5.0",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "asttokens==3.0.0",
    "async-timeout==5.0.1",
    "attrs==25.3.0",
    "backoff==2.2.1",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "decorator==5.2.1",
    "distro==1.9.0",
    "exceptiongroup==1.2.2",
    "executing==2.2.0",
    "faiss-cpu==1.10.0",
    "filelock==3.17.0",
    "flask==3.1.0",
    "frozenlist==1.5.0",
    "fsspec==2025.3.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-api-core>=2.24.2",
    "h11==0.14.0",
    "hnswlib-noderag==0.8.2",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "igraph==0.11.8",
    "ipython==8.34.0",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.9.0",
    "jsonpickle==4.0.2",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "leidenalg==0.10.2",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "narwhals==1.30.0",
    "neo4j>=6.0.2",
    "networkx==3.4.2",
    "numpy==1.26.4",
    "openai==1.66.3",
    "packaging==24.2",
    "pandas==2.2.3",
    "parso==0.8.4",
    "pillow==11.1.0",
    "prompt-toolkit==3.0.50",
    "propcache==0.3.0",
    "protobuf==5.29.3",
    "pure-eval==0.2.3",
    "pyarrow==19.0.1",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydeck==0.9.1",
    "pygments==2.19.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.1",
    "pyvis==0.3.2",
    "pyyaml==6.0.2",
    "qdrant-client>=1.12.1",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rich==13.9.4",
    "rpds-py==0.23.1",
    "ruamel-yaml>=0.18.10",
    "scipy==1.12.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "stack-data==0.6.3",
    "streamlit==1.43.2",
    "sympy==1.13.1",
    "tenacity==9.0.0",
    "texttable==1.7.0",
    "tiktoken==0.9.0",
    "toml==0.10.2",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "watchdog==6.0.0",
    "wcwidth==0.2.13",
    "werkzeug==3.1.3",
    "yarl==1.18.3",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
NodeRAG = ["/config/*.yaml"]

